FM: Sleep timer textview aligned with a Static button.
Previously Sleep Timer view is alligned baseline with RecordMsgTv. On
Honeycomb SDK we figured when RecordMsgTv is not visible even
SleepMsgTv is invisible because of this property set. In order to avoid
this behavior it is alligned with a static button rather.

CRs-Fixed: 284763
(cherry picked from commit 624a54be6a736f353e4a0611f33eea5845b3331a)

Change-Id: Ia7c193e44072ecf06bb2b9c30ec6d80898ef9fe5